Updated Ising Model, added Benchmark function and call from main
